+++ changelog.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- [util] in method: getUniqueConcatenatedValues, also add the promotions for items.buy @maramalshen
 - [actions/***] add 2 keys in integration_meta: sync_to_sap_started & sync_to_sap_succeeded @maramalshen
 - [commands/product] assign Foreign Name (ITEMDESCF) in SAP to local_name in Repzo @maramalshen
 - [actions/payment] add the field UserId where it equals USERID of the SAP API REP @maramalshen